Updated the specs so they work with RJS and add more integrations tests
Updated the specs to work with RJS via selenium-webdriver gem. Using
DatabaseCleaner because selenium isn’t compatible with transactional
fixtures. So this is why we use DatabaseCleaner and manually empty our
database.

Added an extra gem called shoulda-matchers which allows us to do extra
testing on the model like validations, associations and more.